Move InterpreterInformation to pythonEnvironments/types.ts.
