+++ src/client/common/process/types.ts
@@ -1,13 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+
 import { ChildProcess, ExecOptions, SpawnOptions as ChildProcessSpawnOptions } from 'child_process';
 import { Observable } from 'rxjs/Observable';
 import { CancellationToken, Uri } from 'vscode';
 
 import { PythonInterpreter } from '../../interpreter/contracts';
 import { Newable } from '../../ioc/types';
-import { ExecutionInfo, IDisposable, Version } from '../types';
-import { Architecture } from '../utils/platform';
+import { InterpreterInformation } from '../../pythonEnvironments/discovery/types';
+import { ExecutionInfo, IDisposable } from '../types';
 import { EnvironmentVariables } from '../variables/types';
 
 export const IBufferDecoder = Symbol('IBufferDecoder');
@@ -161,16 +162,6 @@ export interface IPythonExecutionFactory {
         resource?: Uri
     ): Promise<IPythonExecutionService | undefined>;
 }
-export type ReleaseLevel = 'alpha' | 'beta' | 'candidate' | 'final' | 'unknown';
-export type PythonVersionInfo = [number, number, number, ReleaseLevel];
-export type InterpreterInformation = {
-    path: string;
-    version?: Version;
-    sysVersion: string;
-    architecture: Architecture;
-    sysPrefix: string;
-    pipEnvWorkspaceFolder?: string;
-};
 export const IPythonExecutionService = Symbol('IPythonExecutionService');
 
 export interface IPythonExecutionService {
