+++ src/client/interpreter/configuration/pythonPathUpdaterService.ts
@@ -2,9 +2,10 @@ import { inject, injectable } from 'inversify';
 import * as path from 'path';
 import { ConfigurationTarget, Uri, window } from 'vscode';
 import { traceError } from '../../common/logger';
-import { InterpreterInformation, IPythonExecutionFactory } from '../../common/process/types';
+import { IPythonExecutionFactory } from '../../common/process/types';
 import { StopWatch } from '../../common/utils/stopWatch';
 import { IServiceContainer } from '../../ioc/types';
+import { InterpreterInformation } from '../../pythonEnvironments/discovery/types';
 import { sendTelemetryEvent } from '../../telemetry';
 import { EventName } from '../../telemetry/constants';
 import { PythonInterpreterTelemetry } from '../../telemetry/types';
