+++ src/client/interpreter/helpers.ts
@@ -4,10 +4,10 @@ import { ConfigurationTarget, Uri } from 'vscode';
 import { IDocumentManager, IWorkspaceService } from '../common/application/types';
 import { traceError } from '../common/logger';
 import { FileSystemPaths } from '../common/platform/fs-paths';
-import { InterpreterInformation, IPythonExecutionFactory } from '../common/process/types';
+import { IPythonExecutionFactory } from '../common/process/types';
 import { IPersistentStateFactory, Resource } from '../common/types';
 import { IServiceContainer } from '../ioc/types';
-import { InterpreterType } from '../pythonEnvironments/discovery/types';
+import { InterpreterInformation, InterpreterType } from '../pythonEnvironments/discovery/types';
 import { IInterpreterHelper, PythonInterpreter, WorkspacePythonPath } from './contracts';
 import { InterpeterHashProviderFactory } from './locators/services/hashProviderFactory';
 import { IInterpreterHashProviderFactory } from './locators/types';
